requirements.txt
absl-py==0.10.0
appdirs==1.4.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorlog==4.4.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
distlib==0.3.1
efficientnet-pytorch==0.7.0
filelock==3.0.12
future==0.18.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.33.1
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
importlib-resources==3.0.0
joblib==0.17.0
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.2
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
Pillow==7.2.0
Polygon3==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.0
torchvision==0.8.0
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.0.28
Werkzeug==1.0.1
zipp==3.4.0
